feat: Refactor providers and update localization keys
