+++ .storybook/preview.tsx
@@ -2,11 +2,11 @@ import '../src/app/globals.css';
 
 import React from 'react';
 
+import Providers from '@/app/[locale]/providers';
 import type { Preview } from '@storybook/react';
 import { Inter } from 'next/font/google';
 
 import enMessages from '../messages/en.json';
-import Providers from '../src/app/providers';
 import { Locale } from '../src/constants/locales';
 import ThemeSwitcherDecorator from './decorators/ThemeSwitcherDecorator';
