+++ messages/en.json
@@ -1,6 +1,6 @@
 {
   "loading": "Loading",
-  "emailOrUsername": "E-mail / Username",
+  "emailOrScreenName": "E-mail / Username",
   "email": "E-mail",
   "username": "Username",
   "name": "Name",
