+++ messages/pl.json
@@ -1,6 +1,6 @@
 {
   "loading": "Ładowanie",
-  "emailOrUsername": "E-mail / Nazwa użytkownika",
+  "emailOrScreenName": "E-mail / Nazwa użytkownika",
   "email": "E-mail",
   "username": "Nazwa użytkownika",
   "name": "Imię",
