+++ src/components/atoms/Checkbox/Checkbox.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import type { ComponentPropsWithRef, RefCallback } from 'react';
 import React from 'react';
