+++ src/components/atoms/SearchBar/SearchBar.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import type { ComponentPropsWithoutRef } from 'react';
 import React, { useRef } from 'react';
