+++ tailwind.config.ts
@@ -173,4 +173,5 @@ export default {
     },
   },
   plugins: [],
+  darkMode: 'class',
 } satisfies Config;
